// src/internal/types.ts
export type RecipeType =
  | "shaped" | "shapeless" | "furnace"
  | "brewing_mix" | "brewing_container" | "unknown";